refactor(l1): make blst the canonical BLS12-381 backend and move the portable fork to the guest program
The BLS12-381 (EIP-2537) trait defaults in ethrex-crypto depended on the
lambdaclass/bls12_381 git fork, which blocks publishing ethrex-crypto to
crates.io. Invert the backend selection: blst is now the default implementation
of the 7 BLS12-381 Crypto methods on the host (the default-on `blst` feature),
and the portable pure-Rust implementation moves into ethrex-guest-program, where
the SP1/RISC0/OpenVM providers override the trait methods with it (Zisk keeps its
FFI path). When blst is compiled out and no provider overrides the methods, the
defaults return the new CryptoError::Unsupported.

ethrex-crypto no longer references the bls12_381 git fork; the fork is pulled
only by the (unpublished) guest program, where SP1's existing patch still
applies via the same git URL.
