+++ crates/common/crypto/Cargo.toml
@@ -25,8 +25,6 @@ ark-bn254 = { version = "0.5.0", default-features = false, features = ["curve"] 
 ark-ec = { version = "0.5.0", default-features = false }
 ark-ff = { version = "0.5.0", default-features = false }
 p256 = { version = "0.13.2", default-features = false, features = ["ecdsa", "arithmetic", "expose-field"] }
-bls12_381 = { git = "https://github.com/lambdaclass/bls12_381", branch = "expose-affine-constructors", default-features = false, features = ["groups", "bits", "pairings", "alloc", "experimental"] }
-ff = { version = "0.13", default-features = false }
 k256 = { version = "0.13.4", default-features = false, features = ["ecdsa"] }
 num-bigint = { version = "0.4.6", default-features = false }
 
@@ -46,8 +44,11 @@ std = [
     "ark-ff/asm",
     "thiserror/std",
 ]
-# Route NativeCrypto's BLS12-381 ops through blst. Native-only (requires std);
-# zkVM guest builds omit it and use the portable bls12_381 trait defaults.
+# Route the BLS12-381 (EIP-2537) trait defaults through blst — the canonical
+# host/L1 backend. Native-only (requires std). zkVM guest builds omit it: the
+# BLS defaults then return an error and the guest `Crypto` providers override
+# them with the portable `bls12_381` implementation that lives in
+# `ethrex-guest-program`, so this crate carries no `bls12_381` git dependency.
 blst = ["dep:blst", "std"]
 secp256k1 = ["dep:secp256k1", "std"]
 c-kzg = ["dep:c-kzg", "c-kzg/std", "c-kzg/ethereum_kzg_settings", "std"]
