+++ crates/guest-program/Cargo.toml
@@ -29,6 +29,11 @@ libssz-derive = { workspace = true, optional = true }
 # zkVM crypto dependencies (pure Rust, compile for RISC-V targets)
 k256 = { workspace = true, optional = true }
 substrate-bn = { version = "0.6.0", optional = true }
+# Portable BLS12-381 (EIP-2537) for the SP1/RISC0/OpenVM guest providers. The
+# git source is fine here — guest-program is not published to crates.io — and
+# matches the URL SP1 patches in bin/sp1/Cargo.toml. Zisk uses FFI instead.
+bls12_381 = { git = "https://github.com/lambdaclass/bls12_381", branch = "expose-affine-constructors", default-features = false, features = ["groups", "bits", "pairings", "alloc", "experimental"], optional = true }
+ff = { version = "0.13", default-features = false, optional = true }
 # Must match ZISK_VERSION in .github/actions/install-zisk/action.yml.
 ziskos = { git = "https://github.com/0xPolygonHermez/zisk.git", tag = "v0.16.1", optional = true, default-features = false, features = ["inputcpy"] }
 
@@ -46,13 +51,13 @@ default = ["secp256k1"]
 
 # Base features: enable the zkVM crypto modules and feature propagation.
 # Guest binaries use these to get their Crypto implementation.
-sp1 = ["dep:k256", "dep:substrate-bn", "ethrex-l2-common/secp256k1", "ethrex-crypto/kzg-rs"]
-risc0 = ["dep:k256", "dep:substrate-bn", "ethrex-l2-common/secp256k1"]
+sp1 = ["dep:k256", "dep:substrate-bn", "dep:bls12_381", "dep:ff", "ethrex-l2-common/secp256k1", "ethrex-crypto/kzg-rs"]
+risc0 = ["dep:k256", "dep:substrate-bn", "dep:bls12_381", "dep:ff", "ethrex-l2-common/secp256k1"]
 zisk = [
     "dep:k256", "dep:substrate-bn", "dep:ziskos",
     "ethrex-common/zisk", "ethrex-vm/zisk", "ethrex-l2-common/zisk",
 ]
-openvm = ["dep:k256", "ethrex-l2-common/secp256k1"]
+openvm = ["dep:k256", "dep:bls12_381", "dep:ff", "ethrex-l2-common/secp256k1"]
 
 # Build-ELF features: base + build tooling for compiling guest ELFs.
 # Used by the prover host. Enabling these triggers the build script which
