+++ crates/guest-program/src/crypto/mod.rs
@@ -3,6 +3,8 @@ pub mod openvm;
 #[cfg(feature = "risc0")]
 pub mod risc0;
 #[cfg(any(feature = "sp1", feature = "risc0", feature = "openvm"))]
+mod bls_portable;
+#[cfg(any(feature = "sp1", feature = "risc0", feature = "openvm"))]
 mod shared;
 #[cfg(feature = "sp1")]
 pub mod sp1;
